requirements.txt
anyio==3.6.1
APScheduler==3.9.1
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.5.18.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
python-telegram-bot==20.0a1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
python-dotenv==0.20.0
typer~=0.4.1
numpy==1.22.4
numpy-financial==1.0.0
dataframe-image==0.1.1
lxml==4.9.0
dateutils==0.6.12
